Move implementation of Component handling into ComponentEntity so that there is only one place the same functionality is defined.
