+++ src/api/libcellml/model.h
@@ -33,136 +33,8 @@ namespace libcellml {
  */
 class LIBCELLML_EXPORT Model: public ComponentEntity
 {
-public:
-    /**
-     * @brief Model Constructor.
-     * Default Model constructor.
-     */
-    Model();
-
-    /**
-     * @brief Model Destructor.
-     * Default Model destructor.
-     */
-    ~Model();
-
-    Model(Model&& rhs); /**< Move constructor */
-    Model(const Model& rhs);/**< Copy constructor */
-    Model& operator=(Model rhs); /**< Assignment operator */
-
-    /**
-     * @brief Add a component to the model.
-     * Adds a copy of the given component to the model.
-     * @param c The component to add.
-     */
-    void addComponent(const Component &c);
-
-    /**
-     * @brief Remove the component at the given index.
-     * Remove the component from the model with the index @p index.  If the index is
-     * not in the range [0, #components) a std::out_of_range exception
-     * will be thrown.
-     * @param index The index of the component to remove (zero-based).
-     */
-    void removeComponent(size_t index);
-
-    /**
-     * @brief Remove the component with the given name.
-     * Remove the first component found that matches the name @p name.  If the
-     * name is not found no component is removed.
-     *
-     * This is an overloaded method.
-     * @param name The name of the component to remove.
-     */
-    void removeComponent(const std::string &name);
-
-    /**
-     * @brief Tests to see if the component is contained within the Model.
-     * Tests to see if the component with the given name 'name' is contained
-     * within the Model.  Returns true if the component is in the Model and
-     * false otherwise.
-     * @param name The name of the component to test for existence in the Model.
-     * @return @c true if the named component is in the model and @c false otherwise.
-     */
-    bool containsComponent(const std::string &name);
-
-    /**
-     * @brief Get a component at index.
-     * Returns a const reference to a component at the index @p index.  If the
-     * index is not valid a std::out_of_range exception is thrown.
-     *
-     * This is an overloaded method.
-     * @param index The index of the component to return (zero-based).
-     * @return A const reference to the component at the given index.
-     */
-    const Component& getComponent(size_t index) const;
-
-    /**
-     * @brief Get a component at index.
-     * Returns a reference to a component at the index @p index.  If the index
-     * is not valid a std::out_of_range exception is thrown.
-     *
-     * This is an overloaded method.
-     * @param index The index of the component to return (zero-based).
-     * @return A reference to the component at the given index.
-     */
-    Component& getComponent(size_t index);
-
-    /**
-     * @brief Take the component at the given index and return it.
-     * Removes the component at the given index position and returns it.
-     * If an invalid index is passed to the method a std::out_of_range
-     * exception is thrown.
-     * @param index The index of the component to take (zero-based).
-     * @return The component at the given index.
-     */
-    Component takeComponent(size_t index);
-
-    /**
-     * @brief Take the component with the given name @p name and return it.
-     * Takes the component at the given index position and returns it.
-     * If an invalid name is passed to the method a std::out_of_range
-     * exception is thrown.
-     *
-     * This is an overloaded method.
-     * @param name The name of the component to take.
-     * @return The component identified with the given name.
-     */
-    Component takeComponent(const std::string &name);
-
-    /**
-     * @brief Replace a component at index.
-     * Replaces the component at index @p index with @p c.  Index must be a valid index
-     * position.
-     * @param index Index of the component to replace (zero-based).
-     * @param c The component to use for replacement.
-     */
-    void replaceComponent(size_t index, const Component& c);
-
-    /**
-     * @brief Replace a component with the given name.
-     * Replaces the component with the given name @p name with @p c.  Name must be a valid name
-     * of a component in the model, if it isn't a std::out_of_range exception is thrown.
-     *
-     * This is an overloaded method.
-     * @param name The name of the component to replace.
-     * @param c The component to use for replacement.
-     */
-    void replaceComponent(const std::string &name, const Component& c);
-
-    /**
-     * @brief Get the number of components in the model.
-     * Returns the number of components the model directly contains.
-     * @return The number of components.
-     */
-    size_t componentCount() const;
-
 private:
     std::string doSerialisation(CELLML_FORMATS format) const;
-    void swap(Model &rhs);
-
-    struct ModelImpl; /**< Forward declaration for pImpl idiom. */
-    ModelImpl *mPimpl; /**< Private member to implementation pointer */
 };
 
 }
