+++ src/model.cpp
@@ -23,62 +23,6 @@ limitations under the License.Some license of other
 
 namespace libcellml {
 
-/**
- * @brief The private implementation for the Model class.
- * This struct is the private implementation struct for the Model class.  Separating
- * the implementation from the definition allows for greater flexibility when
- * distributing the code.
- */
-struct Model::ModelImpl
-{
-    std::vector<Component>::iterator findComponent(const std::string& name);
-    std::vector<Component> mComponents;
-};
-
-std::vector<Component>::iterator Model::ModelImpl::findComponent(const std::string& name)
-{
-    return std::find_if(mComponents.begin(), mComponents.end(),
-                        [=](const Component& c) -> bool {return c.getName() == name;});
-}
-
-// Interface class Model implementation
-Model::Model()
-    : mPimpl(new ModelImpl())
-{
-}
-
-Model::~Model()
-{
-    delete mPimpl;
-}
-
-Model::Model(const Model& rhs)
-    : ComponentEntity(rhs)
-    , mPimpl(new ModelImpl())
-{
-    mPimpl->mComponents = rhs.mPimpl->mComponents;
-}
-
-Model& Model::operator=(Model m)
-{
-    m.swap(*this);
-    return *this;
-}
-
-void Model::swap(Model &rhs)
-{
-    std::swap(mName, rhs.mName);
-    std::swap(mPimpl, rhs.mPimpl);
-}
-
-Model::Model(Model&& rhs)
-    : ComponentEntity()
-    , mPimpl(rhs.mPimpl)
-{
-    mName = std::move(rhs.mName);
-    rhs.mPimpl = nullptr;
-}
-
 std::string Model::doSerialisation(libcellml::CELLML_FORMATS format) const
 {
     std::string repr = "";
@@ -88,85 +32,13 @@ std::string Model::doSerialisation(libcellml::CELLML_FORMATS format) const
             repr += " name=\"" + getName() + "\"";
         }
         repr += ">";
-        for(std::vector<Component>::size_type i = 0; i != mPimpl->mComponents.size(); i++) {
-            repr += mPimpl->mComponents[i].serialise(format);
+        for(size_t i = 0; i < componentCount(); i++) {
+            repr += getComponent(i).serialise(format);
         }
         repr += "</model>";
     }
 
     return repr;
-}
-
-void Model::addComponent(const Component& c)
-{
-    mPimpl->mComponents.push_back(c);
-}
-
-void Model::removeComponent(const std::string& name)
-{
-    auto result = mPimpl->findComponent(name);
-    if(result != mPimpl->mComponents.end()) {
-        mPimpl->mComponents.erase(result);
-    } else {
-        throw std::out_of_range("Named component not found.");
-    }
-}
-
-void Model::removeComponent(size_t index)
-{
-    if (index < mPimpl->mComponents.size()) {
-        mPimpl->mComponents.erase(mPimpl->mComponents.begin() + index);
-    } else {
-        throw std::out_of_range("Index out of range.");
-    }
-}
-
-size_t Model::componentCount() const
-{
-    return mPimpl->mComponents.size();
-}
-
-bool Model::containsComponent(const std::string& name)
-{
-    auto result = mPimpl->findComponent(name);
-    return result != mPimpl->mComponents.end();
-}
-
-Component& Model::getComponent(size_t index)
-{
-    return mPimpl->mComponents.at(index);
-}
-
-const Component& Model::getComponent(size_t index) const
-{
-    return mPimpl->mComponents.at(index);
-}
-
-Component Model::takeComponent(size_t index)
-{
-    Component c = mPimpl->mComponents.at(index);
-    removeComponent(index);
-    return c;
-}
-
-Component Model::takeComponent(const std::string & name)
-{
-    auto result = mPimpl->findComponent(name);
-    size_t index = result - mPimpl->mComponents.begin();
-    return takeComponent(index);
-}
-
-void Model::replaceComponent(size_t index, const Component& c)
-{
-    removeComponent(index);
-    mPimpl->mComponents.insert(mPimpl->mComponents.begin() + index, c);
-}
-
-void Model::replaceComponent(const std::string& name, const Component& c)
-{
-    auto result = mPimpl->findComponent(name);
-    size_t index = result - mPimpl->mComponents.begin();
-    replaceComponent(index, c);
 }
 
 }
